Apply by-design follow-ups to the ttipc 0.1.1 migration
- swap deps to published tauri-typed-ipc 0.1.1 (specta =2.0.0-rc.25), drop taurpc (Cargo + package.json)
- narrow channel_number wire type usize -> u32 (codemod-flagged; exporter rejects BigInt-style ints), cast at usize helper call sites
- add ttipc_bindings() (Snake + createTauRPCProxy drop-in) and a bindings.ts drift test
- regenerate src/bindings.ts
- point the two event-listener hooks at the lifted payload-enum events

+++ src-tauri/src/channel.rs
@@ -10,7 +10,7 @@ pub struct Channel {
     #[specta(type = String)]
     pub id: uuid::Uuid,
     pub disabled: bool,
-    pub channel_number: usize,
+    pub channel_number: u32,
     pub label: String,
     pub label_color: LuxLabelColor,
 }
@@ -28,7 +28,7 @@ impl LuxChannel {
         (*data).clone()
     }
 
-    pub fn get_channel_number(&self) -> usize {
+    pub fn get_channel_number(&self) -> u32 {
         self.0.lock().unwrap().channel_number
     }
